Define the Omnivalidator main menu
Define the menupopup for the "main" menu of Omnivalidator.  This menu
will appear from within the Web Developer item of the Tools menu, the
Web Developer item of the FF 4 App Menu, and the Omnivalidator toolbar
button.  JavaScript will be used to copy the menu hierarchy into these
locations, as there does not appear to be a way to use a single
declaration (short of using XBL).

Also added new translation items to the DTD and new commands for
preferences and showing the about dialog which will be implemented
later.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>