Remove extensionVersion from DTD, consolidate extensionName
Having the version in multiple places only guarantees that one of them
will be overlooked.  Remove it from the DTD.  The button tooltips it was
used for are modified programmatically anyway, so we'll add the version
number in code where required.

Also factor all of the uses of the extension name into a single entity
and reference it in the others.  This should ease translation a little
and make it easier for derivative versions to rename.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>