Allow specifying properties of modules to define
Some files include properties that we are not interested in, which
complicate the usage of the import as a module (particularly when we are
only interested in one property there is no need for the extra object to
wrap it).  Therefore, allow specifying which properties are of interest
when defining a plain or JSM module.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>