Log messages to a file in addition to the console
If users encounter problems and need assistance to diagnose them, it can
be very useful for them to be able to send a log of all the messages
generated by the extension.

Also split the logLevel preferences to allow different levels in the
console and in the log file, with a much lower default level for the
file.

Note:  Overwrite the log at startup to avoid space issues.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>