Introduce "validator IDs" to identify validators
The current system of identifying validators either by their index in
the "array" in the preferences system or by their name, depending on
context, has some annoying flaws.  The major flaw of the preferences
array/index is that when a validator is removed the index of all
subsequent validators changes, making it a poor ID for use outside of
the validator registry and incurring extra cost of deletion to re-index
the array stored in the preferences.  Using the name is flawed because
it may not be safe in all contexts (e.g. it can't be used as a
preference name because it may contain "."), may be locale-dependent in
the future, and it conflates user-centric information with
implementation details.

As a replacement, use a unique identifier (currently a "v" followed by 4
digits) to identify each validator anywhere an identifier is required.

Note:  We could use a GUID for this purpose, but it's unsightly when
appearing in the log messages and preferences due to its length and
inscrutability.  Since we only expect ~10 validators in the average case
and <100 in the pathological case, 5 digits should be more than enough
for now.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>