Rework preferences class/namespace
Make our Preferences API more similar to the underlying implementation.
Particularly, do not treat "." specially except when absolutely
required, and treat the branchName as a filter rather than a base for
listing and observer methods.

Allow the class to be used either statically (which will act on the
root) or as a rooted instance.

Update all uses of the class.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>