Combine defineJSM and definePlain implementations
Most of the code for these two functions is the same, so lets combine
them.

Update error handling for all loaders to log and continue, as would
occur if the file could not be loaded over HTTP.

Additionally, add support for reporting errors when loading optional
modules at a lower severity.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>