Rework automatic validation in validatorregistry
Automatic validation preferences are now stored in the "autoValidate"
value for each validator.  We also only support regular expressions, and
implicitly root them to the start of the URL.  To determine a match,
test the regular expression against the URL (with any fragment removed)
and against the hostname and subsequent portions.

Also remove the obsolete autoValidate default preference.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>