Run all validators on click
For the moment, run all validators when responding to a click event by
the user.  If support for partial validation on click will be supported
in the future, it is likely to be implemented similarly to autoValidate
with expressions determining which validators are run for a click at a
given location.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>