Add id property to Validator
There are places outside of the validator registry where we need a
unique identifier for validators (particularly in the browser validation
manager), which is exactly what the vid is meant for.  So, expose it as
a property of the validator objects.

Make the property immutable to allow clients to depend on the id of a
validator never changing.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>