Factor the console window into a separate class
Move all of the code for the docked console window out of
browser-bootstrap and into its own class ConsoleDockedWin.  Make use of
this class from browser-bootstrap.

Also add detection of location and tab changes to preserve the
visibility state of the window for a given tab/location and to collapse
the window when either the location or selected tab changes.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>