Remove mask from addProgressListener per bug 608628
The mask argument is not supported, as described in Mozilla bug 608628.
However, note that the comments in that bug suggest that omitting
members from the progress listener is supported, while in actuality it
will result in a spew of errors to the error console whenever one of
them would be called.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>