Rename global.jsm to global.js and add global.jsm stub
The purpose of the new stub global.jsm is to be imported using
Components.utils.import to establish a shared global scope for the
extension and to chain-load global.js (either into a sandbox/compartment
for memory debugging, or directly into the shared global scope).

Note:  Use of Components.utils.import within global.js will need to be
replaced in order to account for the memory of the JSMs within the
extension compartment.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>