Add support for status messages to console box
Status messages are special messages which are associated with a key,
such that they can be updated/replaced and removed in addition to the
normal addition/clearing.  They also always appear at the end of the
list of messages.

This will be used to provide validator status while validation is being
performed.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>