Refactor main menu validator loading code
Connect the validate command listener in setupMainMenu and rework
setupMainMenuValidators to reloadMainMenuValidators (which also removes
any existing validators from the menu).  This will allow us to easily
reload the validator list in the main menu when a new validator is added
to the preferences.

Also make use of xulutils for generating the menu list in
reloadMainMenuValidators.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>