Add support for name change listeners in validator registry
Other parts of the system need to be alerted when the list of validator
names changes (e.g. the validate menu must be reloaded).  To support
this, add listener support to the validator registry and notify
listeners when a change in the preference system changes the list of
names.

Note:  This is complicated by validators which lack a name, which still
appear in the list of validators as "Unnamed validator".  For these, we
must detect addition/removal of validators in addition to changes to the
".name" pref.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>