Add to TODO list
It's annoying to not be able to re-run a validator on a page without
reloading.  Although this is not a big use case (since commonly the user
would want to reload the page first), there are situations where it
would be appropriate and should be expected to work (e.g. if the user is
making changes to the validator).

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>