Add deinhibitCaching function to cacheutils
This function adds an observer to the http-modify-request topic which
will remove the INHIBIT_CACHING flag from the loadFlags for document
requests for top-level windows.  This is necessary for the content to be
cached to allow us to validate the content using the cached version.

Note:  This function is currently unused, as I have not found a case
where INHIBIT_CACHING is set on document requests for top-level windows.
However, it represents a significant amount of work and there may still
be undiscovered cases where this occurs, so it is being committed in
case it may be useful later.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>