BIG: Convert validators to use openResourceAsync
Convert both of the validators to use the new openResourceAsync API from
cacheutils.  This requires a bit of extra work, since we need to either
wrap the nsIStreamListener and only send the XHR once we know we can get
the resource, or we need to speculatively start the XHR and abort if the
resource load fails.

For now, we delay starting until the resource is available, since we
want to limit the load on the validator servers as much as possible
(especially for the public validation services).

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>