Remove obsolete functions from cacheutils
We no longer use the synchronous functions from cacheutils or the
function to get the cacheKey for a document (which is done in CacheID).
Since the synchronous functions are incomplete in comparison to the
async functions (they don't support POST or cache failures), get rid of
them.

If we need synchronous functions later, we can pass an
nsISyncStreamListener to the async functions and return the result
(which is what is done internally anyway).

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>