Create XULPrefBranch
XULPrefBranch is an implementation of nsIPrefBranch which stores all
writes in XUL preference elements in a target element and reads
preferences from the XUL with fallback to a next branch.

The main purpose of this class is to work around differences with
systems where browser.preferences.instantApply is false (Windows by
default) where the XUL preferences system stores values in XUL
preference elements until the enclosing prefwindow is accepted.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>