Remove preference state from PrefAVListbox
Rather than storing pending preference changes internally, immediately
store changes in the underlying pref branch.  This allows other clients
of the preference system to see changes immediately and simplifies the
code for the PrefAVListbox.  Also add a constructor parameter for the
prefbranch in which to store the information.

Remove calls to PrefAVListbox.save() in prefwindow as it no longer
exists.  Instantiate PrefAVListbox with an XULPrefBranch for storing
preferences if !instantApply.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>