Implement nsIPrefBranch2 in XULPrefBranch
Implement the addObserver and removeObserver methods required for
nsIPrefBranch2 (and nsIPrefBranch in Gecko >= 13).  These methods listen
for changes either from the next pref branch and from change events in
the DOM tree.

Also rework creation and removal of preference elements to both properly
notify observers of changes and to avoid removing the preference
elements from the DOM, which causes exceptions in the XBL destructor.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>