Move auto VID retrieval out of PrefAVListbox
This function is not actually tied to the implementation of
PrefAVListbox now that it is not storing the preference state
internally.  So, move it into prefwindow, which is the only place that
it is used.

This also allows us to performance-optimize since the URLs are not
required, we don't need to do the queries of the preference system to
get them.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>