Re-register listener in console destructor
The consoleBox XBL expects to be able to unregister the console listener
that it registered in its constructor.  Since we unregistered it in our
subclass constructor, re-register it in the destructor to preserve the
state expected by consoleBox.

Without this, the destructor would throw an exception and fail to finish
completely.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>