Add a width to the prefwindow and persist it
The default width (chosen by sizing of the first prefpane, it appears)
is not wide enough for the automatic validation preferences to be easily
usable.  Therefore, set a default width which should be large enough
without being too large for most resolutions.  Persist it so that on
platforms where users can resize prefwindows (Unix, Mac OS) it is
remembered.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>