Support default branches in NSPreferences
Attempt to present a unified API for Preferences objects where all
methods manipulate the branch they refer to (default or not) with
fallback to default branch for get operations and delete operations.

Some operations (reset) left partially implemented due to the difficulty
of implementing the operation on the default branch (which is currently
not used).

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>