Test/Spec for Preferences with Jasmine
This project is sorely lacking for testing, about time to add some.
Starting with Preferences, which will be implemented for each platform.

These specs can be run in Firefox by creating a page with Jasmine and
loading it from a chrome:// URL.  It's hacky, so I'm hoping for a better
solution before committing it to version control.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>