Standardize preference observer API
Make a Preference API that can be implemented on both Firefox and Chrome
(no weak references, don't pass underlying nsIPrefBranch to observer).
Take a simple callback function which is called with the branch on which
it was registered and the preference name that changed.  Implement for
NSPreferences.

Add specs for the observer functionality.

Update the clients of preference observers to use the new functionality.
Some additional changes required for PrefAVTreeView, which was using
weak references.  Don't assume preferences are being observed and push
enable/disable operations out to clients of the class (who are now also
responsible for disabling preference updates to prevent leaks).

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>