Fix browser (tab) ID calculation
Firefox 15 (ish) added an additional vbox between the browser XUL
element and the notificationbox which has the id attribute that we are
using to identify individual tabs.  This broke our algorithm which only
checked the browser and 2 parent elements.

Rework the algorithm to check all parent elements up to the tabbrowser
element for an id and to warn if the id does not have the format that we
expect.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>